Rename README.txt to README.md; Add README.html to .gitignore
README.md is a WIP and probably will be for a while, as it needs to capture the install and
setup instructions for all the new pieces.  Steps can be added when someone actually works
through that part of the process.

Ignoring README.html.  It gets generated  when I render README.md during editing
to see how it comes out.