Modify Join and Leave RPi messages.
Also, create working but unappealing Station Status page with an Ajax-driven updating table.
Created a 'station_id' field for PiStation that will hold a unique identity that the Pi can use to identify
itself to the MS.  The station_id is passed to the Pi when the Pi Joins.  If the Pi Leaves and re-Joins
with the same info as before, the same station_id will be given out, unless the PiStation record has
been deleted from the database.