Add student Launch Test web page.
Page contains a Google map with colored markers.
Students must use Brata to navigate to specified coordinates, then enter
which marker was reached.

Added a new base template (student_test_base.html) for the student test
pages to inherit from.  This way they will get proper testing URLs and
will not see the administrative URLs.

Added a column to the tests for the Dock test.

Added another value to the Setting table: DOCK_TEST_HOST. It contains the
address of the host running the FlightService for the students to test with.

Changed the URLs in regtest to be relative, so we don't have to update the
server address.

Added some URL patterns.  The test page is reachable as http://<ms>/hsdc/test/