Refactor regtest_team.html and piservice.libraryTest to avoid hard-coded server.
Instead, use reverse() urlresolver and request.get_host() to build the URLs for regtest_team.

That strategy does not work for libraryTest, because the request probably comes from a different
network than the requests that use the QR code will come from.  The QR code needs to contain the
external address of the MS, which may be different from the physical MS address if the MS is
behind a proxy firewall.  LibraryTest uses the MS_EXTERNAL_HOST_ADDRESS value from the Setting
table.